(0x01) feat: getline implementation
Created a simple implementation for the `getline` function. This
implementation will return a line from a file descriptor. If the
function is called various times with the same file, it will return
the next line until eventually, all the contents of the file are read.
In that case the function will return a null pointer.